Fix multihreading issue
Add Lock to prevent multiple threads from accessing some instance
attributes of the socket driver class. This is to ensure that unique
IDs are generated for each command queued in the driver.

The IDs are also cleared and recycled only when the response has been
removed from the responses dictionary. Previously, this was cleared
in the read() method. However, this method is accessed from another
thread (specifically the main thread from the ros service calling the
driver). To synchronise the IDs data, the ID is only cleared in the
get_response() method, which is from the same thread that should be
used to call queue_command() method.

Selector event is also set to READ only mode whenever a command is
popped from the command queue and pushed to driver's send buffer.
The selector event mode is reset back to READ+WRITE only when the
response has been fully received for the last command that was sent
out. This is to prevent WRITE event from occuring in the event loop
and then calling write() which will push the next command in queue to
the driver's send buffer which will also overwrite curr_id and
_find_str which is both needed to be used to receive the full
response for the command that was sent out previously.